+++ packages/polaris-viz/src/components/LineChart/stories/playground/ExternalTooltip.stories.tsx
@@ -63,7 +63,10 @@ const TemplateWithFrame: Story<LineChartProps> = (args: LineChartProps) => {
   return (
     <div style={{overflow: 'hidden', position: 'fixed', inset: 0}}>
       <div style={{height: 100, background: 'black', width: '100%'}}></div>
-      <div style={{overflow: 'auto', height: '100vh'}} ref={setRef}>
+      <div
+        style={{overflow: 'auto', height: 'calc(100vh - 100px)'}}
+        ref={setRef}
+      >
         <Card {...props} />
         <div style={{height: 700, width: 10}} />
         <div style={{display: 'flex', justifyContent: 'space-between'}}>